xlang: Limit hover for didOpen optimization (#5305)
We send a fake hover request when we receive a didOpen notification. We do this
to hide the latency of the initial language analysis a LS does on a file. We
don't just pass on the didOpen since it can mutate state. However, vscode can
send many didOpen requests which do not correspond with a user opening and
browsing a document (eg reference results). We want to keep the performance
benefit of the original approach, but avoid flooding an LS. We accomplish this
by using a rate limiter.