Shutdown unused servers after 30s (#5306)
* xlang: Do not export ShutdownIdleServers

We want to adjust the implementation. All external users of it always passed 0 for
maxIdle, which indicates to kill all servers. So instead we export a function to
do that.

* xlang: Move idle shutdown decision into filter

For example when shutting down everything we don't need to make any decisions,
we can just shutdown. I find this also makes why we are shutting down a server
clearer.

* xlang: Remove idle servers while holding lock

Previously there was a race between deciding to remove an item from the open
server list and actually removing it. This allows another request to find the
open server and send a request to it, even though the proxy is about to or is
shutting the server down.

* xlang: Shutdown unused servers after 30s

vscode reconnection logic means we can open up servers and never use them. This
allows us to more aggressively reap those servers.