searcher: Remove parallelism limiter (#5326)
ParMap limits the parallelism. However, each operation we are doing is just
sending and waiting for a network response, so limiting parallism doesn't help
with perf. Searcher itself is replicated and sharded, so should handle a large
number of queries.

Umami is currently slow to search 160 repos. This change should resolve that.