pkg/vcs: Assume HEAD exists when doing resolve-rev (#5346)
When we do org wide text search, we need to find out HEAD for each repo. When
looking at traces, resolve-rev can be slow (200ms->1.3s). This is an
optimization to increase the speed of resolve-rev. The optimization is to avoid
appending `^0` to the query. This allows git to avoid checking the existance of
the resolved rev. We can only do this for HEAD, since it is the only commit
gaurenteed to exist.