lsp-proxy: Prevent race when creating client conn (#5418)
This was tricky to track down. We occasionally get a panic on `c.conn` being nil
(see attached issue). However, the modified call site is the only place we ever
write to it and it seems this should be safe due to it being initialize code.
However, `jsonrpc2.NewConn` will start processing the handler in the background
which can lead to us accessing `c.conn` before it is actually assigned to.
`c.handle` acquires `c.mu` before doing any work, so by holding `c.mu` when
assigning to creating the connection we will avoid that race condition.