searcher: Avoid parsing new lines if unmatched file (#5419)
This is an optimisation for large repos. It makes our benchmarks for a small repo slower (but still fast overall), and for golang/go it doubles the perf. This change also causes us to use more memory since we allocate upto 2*maxFileSize bytes per search worker. In the future we could optimize that memory usage away by using a sliding window (like bufio works), and or using mmap.