improve defaults for environment vars
This is to simplify the configuration of production and the docker-compose setup (mini). Imho it makes sense to default to the assumption that we have a distributed setup and the services are reachable on their respective DNS names. The dev setup is the special case where everything lives locally, which may also change in the future.