lsp-proxy: Forward $/cancelRequest (#5529)
lsp-proxy multiplexes multiple browser clients onto the same language server. So part of what lsp-proxy does is rewrite JSONRPC2 IDs so that responses can go to the correct browser. Previously the ID sent to a language server would be `$seq:$clientID:$requestID` where:
* `$seq` is a unique number
* `$clientID` identifies the browser client
* `$requestID` was the request ID used by the browser client.

`$seq` was needed since we would automatically retry some requests, so without it we would have non-unique jsonrpc2 ids. However, we have removed that functionality. Without `$seq` given a cancelRequest we can calculate the ID sent to the server as `$params.ID:$clientID`. This PR implements that.