xlang-go: Support for custom GOPATHs (#5674)
* gobuildserver: Factor out detectEnvironment

We want to add more logic to this. It is already quite complicated, so factoring
it out. We also include the related code detectRootImportPath. It doesn't belong
in the root_import_path file.

* simplify stdlib branching

This is a minor cleanup to make later diffs more readable.

* Check all gopaths in findPackageCached

We assume that GOPATH=gopath. However, if we want to allow mono-repo style, we
may have a more complicated GOPATH.

* Handle missing rootImportPath

If we have a custom GOPATH, then we do not set the rootImportPath.

* Detect custom GOPATHs in .vscode/settings.json

Using settings.json was proposed by sqs and is a nice way to tie into what may
already exist in the repo. In the future we could also read from another
location, but we use the same settings that the go extension uses.

* Include GOPATH assertion in test

* Add test for monorepo + GOPATH

* typos