improve vscode JS bundle caching
With this PR, the vscode bundle is now accessed at
`/main/CACHEKEY/...`, not `/main/...`, when running in production.
This `CACHEKEY` is derived from the bundle's VERSION (it is an MD5 sum
to avoid leaking our local usernames, which would just look
unprofessional). When accessed via this path, the Cache-Control header is set to immutable.

This will fix the issue where during a deploy, users could be served 2
bundle files from inconsistent versions. Or they could fail to see the
new version until they hard-reload or until we evict the CloudFlare
cache.