improve correctness of API for repo search endpoint
Previously, there was no way to issue GraphQL repository list queries that retrieved repos existing on GitHub but not existing on Sourcegraph. That should be the default behavior because whether a repo has been mirrored on Sourcegraph is an implementation detail. This led to correctness issues, where Sourcegraph.com's repo list would only show repos that were already mirrored.

This new API introduces a new parameter to repository queries, "fast". The default is false, in which case it queries GitHub to provide a full result set (the correct behavior). Clients can specify fast=true, which currently returns only already mirrored repositories (which doesn't require a call to the GitHub API).

Also, unauthenticated clients are currently forbidden from performing queries that result in a GitHub repo search API call due to the rate limit. This is, unfortunately, an implementation detail that we do need to expose to applications (since we don't control GitHub). Therefore, unauthenticated clients must specify fast=true or else the query will fail.