pkg/localstore: correct Go badge counts
Prior to this change we were not counting distinct repo IDs, which
meant we would over-report on badges the number of users a project
has. The reported number now actually matches the number that will
be found in the list of repos when clicking on a badge (assuming
that list was not limited, which it currently is).