searcher: Log canceled requests differently (#6280)
This will allow us to adjust our alerts and dashboards to correctly reflect that
the error was caused by a user. Currently all our errors in searcher are due to
user cancelation.

It is possibly vscode is canceling requests more often than it should. However,
we should still record this state differently.