cmd/frontend/internal/app: initial streamlined web app implementation
This is very much a WIP codebase, with zero styling/layout performed
yet. This is a good starting point for beginning work on layout and
styling, however. All of the endpoints (repo, tree, blob) have access
to the data that they need and the UI currently shows this in a very
naive way. It's easy to change if we need different data.

- `cmd/frontend/internal/app/templates/ui2` will eventually supersede the templates in the parent directory.
- `cmd/frontend/internal/app/ui2` will eventually supersede the `../ui` directory.
- Adding proper testing for the home/repo/tree/blob endpoints is left as a TODO,
  since they're still very much in flux.

The code will run in production at sourcegraph.com, and can be tested via
setting a cookie:

```
document.cookie="streamlined=true;path=/"
```

And likewise you can get the regular UI back via:

```
document.cookie="streamlined=false;path=/"
```