cmd/frontend/internal/graphqlbackend: separate listing active repos from search backend
This is a follow-up to https://github.com/sourcegraph/sourcegraph/pull/6499 in which
Keegan had some very good feedback about the fact that we should separate the listing
of active/inactive repos to a different graphql root endpoint, and then cache that list
on the client effectively (and that it is also useful for other things like repo
switching etc. in the future).