webapp: correct window.pageVars.Rev to not contain @ prefix
This was a regression caused by my larger URL generation refactor
last night, and the problem was that `window.pageVars.Rev` unexpectedly
contained an `@` prefix in it. This would lead to a double `@@rev` in
URLs as seen in #6548.

I don't expect any other regressions to have occurred, since the refactor
was primarily about making the URL generation code more understandable when
adding revisions to URLs.

Fixes #6548