update repo via GitHub API on a common code path
Fixes #6278. The previous code path it not used much any more.

The comment that got removed was not valid. Our security model does not depend on the "private" field to be updated "in time". As long as our system considers a repo public, no new private commits get fetched, thus not exposing any data that wasn't public at some point.