gitserver: Do not hold cloning lock while running git ls-remote (#6757)
Previously we would do a network request while holding the lock. The negative
effect of this is noticeable when doing global search on uncloned repos on a
poor network connection. I also think this is a good explanation for the strange
brief slowdowns we occasionally see in gitserver for all commands. This is
because the cloningMu has to be acquired before running any command.

This also has the nice side-effect of moving the unlocks next to the locks on
the mutex.