webapp: ensure tooltip search uses current search state, not localstorage state
Prior to this change, searching from a tooltip would use the localstorage search state
(for example, if you previously entered "github.com/foobar/baz" on the homepage, it
would use that). The expected behavior is the same as searching from a repo page: that
it would search inside just that repository. This change effectively does that.

Fixes #6754