web: make blame info open relevant commit page
This change makes it so that clicking on the blame text of a code line opens the
GitHub commit view. As the inline comments say, the implementation is not perfect
from a code quality standpoint, and I'd like to re-implement blame in a better way
in the future, but for now I think this is good enough to merge because it is a
clear improvement for users and has no user-facing drawbacks.

This helps #6544 (notably, that issue includes Umami support and this doesn't
have it yet).