pkg/highlight: use custom Sourcegraph theme
This change causes us to use a custom Sourcegraph theme whose colors are based on
existing designs by Leo and our existing syntax highlighting theme in `Sourcegraph - Insiders`.

We can adjust the colors further in the future as we please, but for now the primary
goal is that this PR:

- Fixes #6931
- Fixes #7150

The theme is originally based on `base16 ocean (dark)` as that theme produces the best (most
distinguished / etc) results in multiple code files.

You can stop reading here if you please. :)

This venture has shown me that, unfortunately, due to the way TextMate syntaxes and color themes work:

1. A TextMate/Sublime syntax file has some regexp which matches a section of code and gives it an arbitrary
   name like `variable.parameter.function` which has no true formal standard.
2. A color theme lists base background and text color, and then specifies arbitrary keys matching the syntax
   files like `variable.parameter.function` to apply a color to it.

The theme files are inherently linked to the syntax definitions for languages. There is some commonality
between them, which is why the theme I based ours on (`base16 ocean (dark)`) works better than others, but
for example the TypeScript syntax file emits keys that are not specially colored by our theme, but could be.

We would have to inspect all language syntax files in order to achieve perfect results. In the future, we
could list some base colors and use a generation script to do the grunt work for us, which would give us
some pretty flawless themes.

https://www.useloom.com/share/e31fca96449a4868965074784749eaff

Compare with `Sourcegraph - Insiders`:

![image](https://user-images.githubusercontent.com/3173176/30445876-548669ac-993c-11e7-9ad8-d89d6595d8fe.png)

Fixes #6931 (remaining issues there are covered by #6932 )

https://www.useloom.com/share/1febc5adce5548f09fbb28f178dcc32f

Compare with `Sourcegraph - Insiders`:

![image](https://user-images.githubusercontent.com/3173176/30446250-68aa6194-993d-11e7-8fed-1122e4042b69.png)

https://www.useloom.com/share/de1aa235f9ab4793a44f97b3d55c82fd

https://www.useloom.com/share/7ebaaba6252344609b78684de83f994d

Fixes https://github.com/sourcegraph/sourcegraph/issues/7150