disable omnisearch for now
For now, disable omnisearch. If we have time, we'll add it back according
to one of the approaches found in [Felix's comment here](https://github.com/sourcegraph/sourcegraph/issues/6798#issuecomment-329325560).

I think it is unlikely that most users will be using omnisearch right now,
and therefor would like to say it's OK if it doesn't get done by the 20th
(but that we should fix soon, still).