web: remove faulty redirect from repo to search (rebase/merge error)
Not sure how I ended up getting this code here, but it's not correct.
A repo page should never redirect to `/search`, except in the case of
a legacy query param (which is already handled above explicitly).

This must have been a rebase/merge error, or a copy-pasta error.

This is now the correct behavior, to render the SPA for repo pages.

Fixes #7219
Fixes #7231