Repos.List is less fuzzy (#7290)
It now breaks a query up into alphanumeric runs, and searches for those. It
relies on higher layers to further filter out matches and rank them.

The previous implementation was too fuzzy and lead to poor performance. This
implementation should allow postgres to use its trigram index for most queries.