Rank repos assuming query is matching from end (#7327)
This changes the ranking of repo results for a query. We now always award
the straightforward `Score(repo, query)`, but additionally award a higher
score for queries that align well with the suffix of a repo, split by path
components. This works from the assumption that if your query is `foo`, you
should award a higher score for matches to `foo` in the basename of the repo
uri (ie the repo name in github). If you change your query to `bar/foo`, then
you should award a higher score for `foo` matching in the basename, but
additionally award a higher score for `bar` matching in the next basename (ie
the org name in github). Previously we would award points for component
matches anywhere. I believe this more closely matches user's intent.

A second less important change is that we don't throw away separator
characters when querying the repo table. Previously we would just return many
results we would just throw away at the scoring stage. In theory this also
makes the query faster, since there is more "ngrams" to filter on.