web: fix line highlighting / blame on files that do not have language support
Prior to this change we would only allow clicking on / highlighting lines of
files that are of a language we do not support when you clicked on a subsequent
file and back.

For example:

1. Load https://sourcegraph.com/github.com/linkedin/LayoutKit/-/blob/Sources/Layouts/OverlayLayout.swift
2. Clicking lines / blame does not work
2. Click a different file, click back to same file.
3. Clicking lines / blame now works.

This fixes the issue by ensuring we properly register event handlers on initial page load.

Fixes #7395