graphql: sanitize Thread.html* before returning to client
Prior to this change, all HTML from `Threads.htmlBefore`, `Thread.html`, and
`Thread.htmlAfter` was unsanitized. Primarily because nobody is actually
putting information in these fields yet.

We can't trust the editor to send sanitized HTML (obviously), and storing the
sanitized HTML is not good either because it would be lossy. So with this
change, we continue storing the unsanitized user input and sanitize it before
returning it to the graphql caller.

This change also clarifies at each location (graphql schema and localstore
backend API) whether or not the user of the field can expect the HTML to be
sanitized or not.

The very popular bluemonday sanitizer is used, with it's UCGPolicy, and a
single (very well tested) exception to let the syntax highlighted code
generated by VS Code pass through the `style="color: #123123;"` tags on the
spans that it generates.

Helps #7517 and #7516