backend: prepare shared item URLs to be public (no behavior change)
The goal of this PR is to support _two_ different security models for shared
item URLs:

1. The author of the shared item only wants the shared item viewable by members
   of their organization (this is what we have today).
2. The author of the shared item wants anyone with the link to be able to read
   the code snippet and the entire discussion thread, and be able to reply from
   the web if they are signed in (so they have a username), even if they are
   not a part of the org.

Security model 2 above (secret URLs) is the primary model that all users will
make use of. In fact, we won't even offer the "only members of my org" security
model to users for a while (i.e. until a customer asks us for it).

Security model 1 above is useful because we want to generate shared URLs for
all email and slack notifications, and having these URLs be completely public
would be very concerning. e.g. if you share a link you receive in a Sourcegraph
email with the wrong person, they could see the discussion and even reply. So
these links will be restricted to organization members only.

This change prepares us to support the second security model above by
restricting which fields a user with a 'secret URL' (security model 2 above)
can query over graphql. For example, just because you have a valid secret URL
does not mean you should be able to query graphql `root.sharedItem.thread.repo.threads`
and read every single thread in that organization.

To do this we alter the graphql schema such that shared items only expose a
safe subset of the information that we believe users are OK exposing when they
share a secret URL.

Left as a TODO is to make the editor create secret URLs by default (security
model 2), and to store this as a `public` boolean field in the database. I've
intentionally left this out of this PR because this PR is _just_ about how to
support the two security models.

Helps sourcegraph/src#1458
Helps #7736