cmd/frontend/internal/app/ui2: fix editor/browser ext search redirect
With the new search UX PR from @sqs, we modified this file in two ways:

1. To remove the redirect.
2. To remove the title setting.

Removing the title setting ([see discussion here](https://github.com/sourcegraph/sourcegraph/pull/7749/files#r146759439)) was an accident.

The argument for removing the redirect was that it has been around
long enough (I am a bad reviewer and forgot to verify this), and that
it was hard to keep around because we couldn't check the user localStorage
for the new feature flag on the backend.

Removing this was a mistake for two reasons:

1. Both the browser and editor extensions have not been updated to use
   the new search URLs, and so they have been broken since we merged
   this change/
2. Checking the feature flag should not be needed here, since it is a
   hard redirect. They land on the right page regardless, and then the
   feature flag takes over.

Fixes #7822