bypass old query URL redirect if using new query UX URLs
Addresses the issue that necessitated the hotfix in https://github.com/sourcegraph/sourcegraph/pull/7824.

As explained in https://github.com/sourcegraph/sourcegraph/pull/7824#issuecomment-340638273, we need to bypass the old redirect from any repo page with `?q=` to the search results page. That interferes with the expected behavior of the new query UX, which is to maintain the `q` and `sq` URL params as you navigate, if you had come from a search results page.

The way we usually identify whether you're in new query UX is with `localStorage.search2`. This is not available to the server, so instead we ensure that new query UX URLs always contain a (possibly empty) `sq` URL query param, and we check for that before redirecting the old URL route.