preserve 'q' and 'sq' url query params when clicking line in blob view - fixes #7837
Previously, when we appended the `#L123`, we would remove the 'q' and 'sq' URL query params for search. This change makes it so they are preserved when you click on a line.

Watching component updates (props changes) is a more reliable way than listening to the router to react to URL changes, because it is called before the router listeners. The router listener sometimes was called after this.props had changed, which meant it was not possible to determine the values of 'q' and 'sq' prior to the URL change.