fix tree bug when expanding an entry with a same-prefix sibling
Fixes a bug in the file tree where expanding `abc` in the following tree would cause both `ab` and `abc` to be expanded:

```
mydir
  ab
  abc
```

This was caused by incorrect logic to check if `mydir/ab` should be expanded. It was checking whether `mydir/ab` was a string prefix of the currently expanded path. Instead of checking for string prefix, it needs to check whether it's an ancestor dir. I.e., `mydir/ab` is a string prefix of `mydir/abc` but it is NOT an ancestor dir of `mydir/abc`.

This change fixes it to use ancestor path logic.

Live repro: expand pkg/apimachinery at https://sourcegraph.com/github.com/kubernetes/kubernetes and notice that the pkg/api directory is expanded too.