pkg/localstore: correctly fetch public/non-public shared item before returning (#7860)
When creating a shared item, the existing one is returned instead of creating
a new one (only if one already exists).

But this was not based on whether or not the shared item is public or not. For
example, trying to create a public shared item right now actually creates an
organization-only shared item because notifications are first sent out for all
comments (which share the item as org-only).

This fixes that issue by ensuring we only return the existing shared item if it
truly matches the item we're trying to create.