only redirect / to about site on sourcegraph.com
On dev and staging servers, it is a slight inconvenience to have to remember to visit http://HOST/search to see the actual site, because the root / path redirects to about.sourcegraph.com. This commit makes it so only sourcegraph.com/ and www.sourcegraph.com/ redirect to the about site.

On-prem deployments are already exempt, so they are not affected by this change.