redirect old (pre-streamlined web app) search URLs to new query UX URLs - fixes #7976 fixes #7975
Previously this redirected `/myrepo?q=foo` to `/search?q=foo&repo=myrepo`. That was the URL for the streamlined web app search before the new query UX. The redirected-to URLs would not work if the new query UX was enabled on the server (which is the default for sourcegraph.com and all Server instances).

Now it redirects `/myrepo?q=foo` to `/search?sq=repo:^myrepo$&q=foo`.

This was only a problem for Firefox and the editor extensions AFAIK, not Chrome.