redirect /server to about.sourcegraph.com/products/server (#8011)
* allow specifying a map for redirects to about.sourcegraph.com

This commit does not change behavior. It is a refactor to prepare for the next commit, which will introduce a redirect from one path on the server to a *different* path on about.sourcegraph.com.

* redirect /server to about.sourcegraph.com/products/server

I want to be able to tell people verbally "go to https://sourcegraph.com/server" instead of "go to https://about.sourcegraph.com/products/server" (which is longer).