improve search filtering of specific repo revisions (fix #7851) (#8032)
* search: make 'repo:golang/go' be implicitly treated as 'repo:(^|/)golang/go' (helps #7851)

* search2: improve searching for specific repo revisions (fixes #7851)

Previously, if you searched for a specific repo revision and it matched multiple
repositories, it would fail unless that revision existed in both repos:

![image](https://user-images.githubusercontent.com/3173176/32685434-6ef67c0c-c64e-11e7-96e3-031b5071a231.png)

Now, we no longer fail in this situation. Instead, we present the user with the
repositories that do contain that revision, and warn the user that:

> A repository matched by your repo: filter could not be searched because it does not contain the revision "go1.8.3".

i.e. instead of the above screenshot, you now get:

![image](https://user-images.githubusercontent.com/3173176/32685437-83c7be7a-c64e-11e7-8952-7269bd00528e.png)

* search2: update tests to reflect new "(^|/)" prefix of "repo:"

* apply repo:pattern@rev to repos that match pattern case-insensitively

Previously, if you had repo:foo and it matched a repo named Foo because the backend.Repos.List call was case-insensitive, the getRevForMatchedRepo would not match it because there the regexp was case sensitive. This fixes that.

* say which repo could not be searched

* remove debug log