deep link to branch || repoRevision on comments page (fixes #7989) (#8042)
* remove revision from comments page deep links (fixes #7989)

* web: deep link to branch || repoRevision

On further testing I found deep linking to no revision does not change the
branch listed in the SCM tab, in contrast deep linking to an exact revision
does. This means when we deep link to a branch that no longer exists, it will
fail currently (because the editor does not handle this gracefully and lookup
the revision from the thread ID), but we'll be changing this in the future
anyway so it doesn't matter.