pass through context deadline and canceled errors instead of treating them as ErrRepoNotFound
Prior to this change, if the caller used context cancelation or timeouts and one of those occurs while executing this function, the error that's passed back would be ErrRepoNotFound. It returns that error in all cases because it does not want to leak any other error value that might reveal sensitive info. However, it makes it so callers can't (correctly) suppress errors that were caused by cancelation/deadlines.

This change makes it pass through those 2 specific error values. This is secure because they do not leak any sensitive info.