simplify search results header - remove # files/# repos
Files and repos are concepts that do not apply to (or are not the ideal way of counting) all of the search result types we want. Also, # repos was misleading because many users thought it meant that's how many repos were in scope, not how many had results.

Also makes the font smaller and padding tighter, to save vertical space (#7995).