searcher: Open a reader per request and stream directory headers (#8045)
* govendor add github.com/sourcegraph/lazyzip

* vfsutil: Factor out pkg diskcache

* searcher: Use diskcache and lazyzip.Reader

This greatly simplifies our Store code since we no longer have to keep track of
open readers. Instead we open a reader for every request. This is better since
the size of the `File` slices in a `zip.Reader` was getting so large that our
service would OOM on startup.

The diskcache package has the same behaviour as our old store code w.r.t. single
flighting and storing on disk. We also get to simplify the LRU implementation
and just evict items based on the file modtime.

* vfsutil: Add back test utility useEmptyArchiveCacheDir

* searcher: Remove race condition on err