web: fix regression that broke external global references (fixes #8099)
A recent commit (cc @felixfbecker) entirely broke external global references.
The faulty change is:

https://github.com/sourcegraph/sourcegraph/commit/0a84848befbd6e44198b636e0ac0a232d3c13958#diff-f68b248367ab677c30bf13232bb7fa07L360

Prior to this change we did:

```
return json[1].result.map((data: ReferenceInformation) => data.reference)
```

Where `data.reference` is the `Location`. But after this change, we do not do
this and instead just return the result directly. That is, despite the signature
of the function being `Promise<Location[]>` the function is actually returning
`Promise<ReferenceInformation[]>` and the caller trying to access the `uri` field
ends up with `undefined` when they should not.

Why didn't our TypeScript compiler pick up this type error? I am not sure.
I will defer the correct long-term fix here to @felixfbecker since I have a
feeling he has a better idea of how to ensure that we have type correctness
here.

Fixes #8099