add saved queries, persisted in user or org config
Saved queries let you save a search query so you can return to it easily. Users and orgs with the `saved-queries` tag enabled will see a new "Queries" navbar link. This page supports creation, editing, and deletion of saved queries, and it shows a summary of the saved queries' result counts.

The goal of this feature is to (1) make it easy to define queries that match code patterns that you'd like to be flagged during code review in the checklist viewlet and (2) provide basic reporting functionality for engineering managers (related to (1) but even if they are not using the editor).