show signout as button
After the previous commit made it so non-editor-beta users would still see orgs, the signout button looked weird because there was no divider between the orgs and the signout button. This makes it into a true button and adds a divider, using the same style as for the "Download Editor" button.

The two possible permutations are:

In editor beta:

![image](https://user-images.githubusercontent.com/1976/33305596-bb87a424-d3c3-11e7-9d6c-621b7fa068a7.png)

Not in editor beta:

![image](https://user-images.githubusercontent.com/1976/33305611-cbc3dcfe-d3c3-11e7-8051-7830a1392a7c.png)