limit sending org invites by user, based on a quota, to prevent invite abuse
This will let us allow all users to create and join orgs on Sourcegraph.com (instead of limiting it to editor-beta users as is the case now). This means all users can use shared org config for saved queries and search scopes. Allowing that is the motivation.

Applies to non-on-prem instances only (of which Sourcegraph.com is the only one), NOT Sourcegraph Server on-prem instances.

A Sourcegraph.com user who invites more than 15 (hardcoded default PostgreSQL column value) users across all orgs is shown an error that they are over the invite quota and they need to contact support. We then need to manually increase their `invite_quota` value in the DB to lift the limitation.

Users will see this when they are over quota:

![image](https://user-images.githubusercontent.com/1976/33308046-c70611a0-d3cd-11e7-8585-6004fda41268.png)