localstore: Remove Repo.DefaultBranch (#8164)
* localstore: Remove Repo.DefaultBranch

It isn't gaurenteed to be set, which lead to several bugs when it isn't. Note
that we still expose DefaultBranch in graphql, but that value will always be
computed via gitserver.

* graphql: Mark DefaultBranch as optional

If we are cloning the repo, DefaultBranch field will fail. Rather than fail the
whole graphql request, just do not return the field.