make full-width take up 100% width on all screen sizes
Previously above media-md, it had 5rem horizontal padding, which was a meaningful amount of space that was
unusable. This change makes it so that the media-md behavior (100% width) is used for all
screen widths.

This means the file tree toggle must always be shown on the right. Previously it covered
up the line numbers; this change makes the toggle smaller (and have a tooltip for easier
discoverability) to avoid that.