auth.userOrgMap server config: enforce all users are members of specified orgs
This is useful for customers that want to use orgs but want to have everyone be a member of at least one org automatically (which is everyone). This will make it possible to configure search scopes, saved queries, and more future configurable things for all users on the server, without needing to drop down to infra-level config (org members can just edit the JSON file on the web).

Documented in https://github.com/sourcegraph/website/pull/100, which should be merged AFTER this PR is merged