web: comments: support rendering of plaintext comments (fixes #7773) (#8311)
Prior to this commit, if you shared a snippet or comment that was made in a
plaintext file (i.e. one that was not highlighted in your Sourcegraph editor)
then it would not render any code on the sourcegraph.com/c/... page. This change
fixes that by properly rendering the plain text lines.