web: markdown: move max-width into child container
Inside of https://github.com/sourcegraph/sourcegraph/commit/bacab1dd210d1a342c98911a80b953c30fab6406
I moved the max-width property into the `.repository__rich-content` container so that
it did not restrict the width of the markdown on comment pages, which was quite ugly.

However this had the side effect of causing the scroll bar on Markdown pages to be shown in the wrong
location because `.repository__rich-content` must have `overflow: auto;` on it. So it causes it to
look like:

![image](https://user-images.githubusercontent.com/3173176/33586621-017930fa-d91f-11e7-8e43-5b73103e753b.png)

This obviously isn't good, so we need the max-width to be on the actual child
content and not the scrolling container. This means I can either use a
`.repository__rich-content .markdown` selector which violates BEM or introduce
a new container element; so I've opted for the latter.