backend: make comments email notifications have proper line break (fixes #7927) (#8312)
This change fixes #7927 by ensuring we always correctly join the text lines that
we send as part of our email notifications.

The fields (`thread.Lines.TextBefore`, `thread.Lines.Text`, `thread.Lines.TextAfter`)
do not have trailing newlines mainly because it significantly complicates the editor
code which captures these lines. Joining them with a newline is trivial, and so we do
that. I works in all cases (I've tested / confirmed it).

Fixes #7927