xlang: Check if commit exists before forwarding request (#8448)
Before this change we could establish a connection to a new language server and
not actually have the commit. This would lead to it failing later when the
language server actually attempted to access a file. Its better to return the
error sooner since the other errors are quite hard to understand.