Don't hang on SSH host key prompt when communicating with SSH repos on gitserver
fix #8499

If gitserver tried to run a command that communicated with a remote repo over SSH, it would prompt the user 'yes/no' to add the server's SSH host key to known_hosts if it had not yet seen the host key. This would cause the gitserver goroutine to hang indefinitely because the command was not being run interactively and the prompt would be unanswered forever. (Strangely, the git process also seemed immune to being killed without -9.)

This uses the `GIT_SSH_COMMAND` env var introduced in git 2.3 (2015 Feb 6) to set the `-o StrictHostKeyChecking=yes` option for `ssh`. That option causes `ssh` to immediately fail instead of asking the user yes/no to add a new host key to known_hosts. This fixes the issue.

With this change, gitserver will report errors if asked to communicate with host-key-unrecognized remote repos:

```
19:29:16        indexer | t=2017-12-14T19:29:16-0800 lvl=warn msg="Could not ensure repository cloned" uri=35.197.32.22/renfred/holiday error="repository does not exist"
```

Perhaps this error should be nicer, but we already generally translate any auth error into "repository does not exist", so this is fine for now.

To make these repos actually cloneable, the user needs to add their hosts' host keys to known_hosts.

This is strictly more strict than before, so there are no known security issues introduced with this change.