add locally unique `key` field to saved queries
This lets us refer to a specific saved query elsewhere (e.g., the upcoming saved query alerts feature) in a way that is robust to the saved query's index in the `search.savedQueries` array changing.

All existing saved queries are migrated to have a `key` field added when they are next fetched.

We use configuration, not the DB, to store saved queries for simpler cascading/merging (of orgs + repos + user configs), and editing of advanced config using JSON schema instead of needing to design UIs for everything. However, some features can't be fully expressed in JSON config (such as whether a user has agreed to receive alerts for a saved query), so we do need a way to refer to config items externally (using this `key` field in this case).