remove no-longer-needed user backfill
The user backfill code was for the switchover to the streamlined web app. It let users who were logged in before the switchover more easily create an account in the new Auth0 system. It has been a few months and now the number of users who it would apply to is extremely low (most users would have either visited since then, or their cookie would have been expired/deleted).

This is a nice reduction in complexity of the code now that we don't need the backfill anymore.