use Monaco editor for user/org JSON settings (feature-flagged OFF by default)
Run `localStorage.monacoSettingsEditor=true` in the JS console and reload to use the Monaco editor on user/org profile pages for the settings JSON editor.

See known issues in MonacoSettingsEditor.tsx TODO list. These will need to be addressed before enabling this for all users (not behind a feature flag).