use `repositories: RepositoryConnection!` in graphql API to allow for totalCount field
The GitHub API uses `XyzConnection` types as the result type for many list operations. This `XyzConnection` type contains information about the relationship between the subject and the item, plus it supports pagination and counting-without-fetching.

This commit introduces a basic `RepositoryConnection` that is modeled on GitHub's.

All call sites are updated (none exist outside this repo). The docs were already updated to just use `repository(uri: ...)` so they are unaffected by this change.